requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.15.0
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-cloudinary-storage==0.2.3
django-cors-headers==2.5.2
django-import-export==1.2.0
django-proxy==1.1.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.4
mock==2.0.0
odfpy==1.4.0
openpyxl==2.6.0
pbr==5.1.3
psycopg2==2.7.7
python-dotenv==0.10.1
pytz==2018.9
PyYAML==4.2b1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
